pcap_replay_capture/requirements.txt
tdl-xoa-driver>=1.2.0
scapy>=2.5.0